add custom condition to end workflow
